Add setup.py for installation, move code to src subfolder.
